right before destroying

+++ app/src/main/java/com/example/basketballorangrybirds/GameActivity.java
@@ -21,7 +21,7 @@ public class GameActivity extends AppCompatActivity {
         Point point = new Point();
         getWindowManager().getDefaultDisplay().getSize(point);
 
-        gameView = new GameView(this, point.x, point.y);
+        gameView = new GameView(this, (short) point.x, (short) point.y);
 
         setContentView(gameView);
